# ARROW-7526: [C++][Compute] Optimize small integer sorting