--- a/cpp/src/arrow/compute/kernels/sort_to_indices_benchmark.cc
+++ b/cpp/src/arrow/compute/kernels/sort_to_indices_benchmark.cc
@@ -49,10 +49,31 @@ static void SortToIndicesInt64(benchmark::State& state) {
   SortToIndicesBenchmark(state, values);
 }
 
+static void SortToIndicesInt8(benchmark::State& state) {
+  RegressionArgs args(state);
+
+  const int64_t array_size = args.size / sizeof(int8_t);
+  auto rand = random::RandomArrayGenerator(kSeed);
+
+  auto values = rand.Int8(array_size, -100, 100, args.null_proportion);
+
+  SortToIndicesBenchmark(state, values);
+}
+
 BENCHMARK(SortToIndicesInt64)
     ->Apply(RegressionSetArgs)
     ->Args({1 << 20, 1})
     ->Args({1 << 23, 1})
+    ->MinTime(1.0)
+    ->Unit(benchmark::TimeUnit::kNanosecond);
+
+BENCHMARK(SortToIndicesInt8)
+    ->Apply(RegressionSetArgs)
+    ->Args({1 << 20, 1})
+    ->Args({1 << 23, 1})
+    ->Args({1 << 23, 50})
+    ->Args({1 << 23, 80})
+    ->Args({1 << 23, 99})
     ->MinTime(1.0)
     ->Unit(benchmark::TimeUnit::kNanosecond);
 }  // namespace compute