@wesm 2020-01-13
Thanks! Note that this approach also yields perf improvements for "small range" sorts on larger integer types, even when you add in the cost of computing the range

see https://issues.apache.org/jira/browse/ARROW-1571

Perhaps someone will explore this work in the future